requirements.txt
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0
tensorboard>=2.16.2
wheel>=0.43.0
typing_extensions>=4.11.0
aiolimiter>=1.1.0
matplotlib>=3.8.4
optree>=0.11.0
pandas>=2.2.2
python-dotenv>=1.0.1
PyYAML>=6.0.1
Requests>=2.31.0
rich>=13.7.1
numpy>=1.26.4
scipy>=1.13.0
tqdm>=4.66.2
click>=8.1.7
notebook>=7.1.3
ipywidgets>=8.1.2
jupyterlab>=4.1.6
wandb>=0.16.6
streamlit>=1.33.0
joblib>=1.4.0
tiktoken>=0.6.0
openai>=1.23.3
deepspeed>=0.14.2
accelerate>=0.29.3
huggingface-hub>=0.30.2
datasets>=2.19.0
peft>=0.10.0
transformers>=4.40.1
trl>=0.8.6
sentencepiece>=0.2.0